refactor(docker): fill asm-params via one-shot compose service
