+++ docker/README.md
@@ -148,7 +148,9 @@ done
 
 Use that output to populate `docker/vol/strata-bridge-{1,2,3}/params.toml`.
 
-The ASM runner uses the same operator set by reading `docker/vol/strata-bridge-3/params.toml` at startup.
+The ASM runner's operator set lives in `docker/vol/asm-runner/asm-params-sample.json` and must
+be kept consistent with the bridge `params.toml` by hand — the runner no longer reads
+`params.toml` at startup.
 
 ### ASM Runner
 
@@ -159,18 +161,16 @@ The ASM runner is part of the Docker stack and is exposed on:
 The relevant ASM files are:
 
 - `docker/vol/asm-runner/config.toml`: checked-in runtime config
-- `docker/vol/asm-runner/asm-params-sample.json`: checked-in reference/sample params
+- `docker/vol/asm-runner/asm-params-sample.json`: checked-in **partial** params; the input template for the fill step below
+- `docker/vol/asm-runner/asm-params.json`: generated by the fill step, gitignored, mounted at `/app` and consumed by the container as-is
 - `docker/vol/asm-runner/data`: persisted ASM database
 
-The sample params file is documentation only. It is not used directly at runtime.
+The asm-runner expects `asm-params.json` to already be present in `/app` with anchor data that
+is consistent with the L1 it will be running against.
 
-At container startup, `docker/asm-runner/entrypoint.sh` runs `docker/asm-runner/gen_asm_params.py`, which:
-
-1. reads `docker/vol/asm-runner/config.toml`
-2. reads bridge operator MuSig2 keys from `/bridge/params.toml`
-3. queries the live regtest bitcoind node
-4. writes an ephemeral params file to `/tmp/asm-params.json`
-5. starts `strata-asm-runner --config /app/config.toml --params /tmp/asm-params.json`
+For the docker compose stack, the one-shot `asm-params-init` service does the fill against the
+internal `bitcoind` before asm-runner starts, so `docker compose up` works end-to-end with no
+host steps.
 
 Bridge nodes talk to ASM over Docker networking using `http://asm-runner:9010`.
